Day 06
Part A was suspicously easy, but apparently a trap for part B. Since
I used recursion for part a - which seemed obvious - it would easily
blow out of proportion with larger day counts and hit memory limits etc.

So I thought of a different approach:

Basically grouping the days-until-birth into an array with length 8.
Each step I'd take the `head` - which is the number of new fish this
iteration. And add that amount to the end (8 days) and add it to the
amount at 6 days.

Much faster and memory efficient than recursion.

This was a great one!