🦀 Day 07
This was just plain fun.

Initially I misread the text/assignment and thought it was way more complex than what as actually asked for. When I started to implement it, I was like "wait, or is THAT already it?" and it was and finished it up.

I was weirdly excited about Part B, because you rarely get the chance to implement factorial and since it seemed like a similar memory/performance trap to Day 06. So I memoized the factorial function directly. Only tried running without memoization afterwards, waited a bit but didn't even let it finish.

Some of the challenges feel like they're made with function composition in mind. I think this was the third time already where I only had to pass in the core-calculation function to pass part B (identity vs factorial)

Deciding to include Ramda was a great idea. My fp helpers looked like the Ramda API anyway. But I'm glad I started building the core fp-helpers myself.